Rename nonlinear solver and namespace
